requirements.txt
atakama>=1.7.1
python-dateutil~=2.8.2
docmd
pytest~=7.0.1
coverage>=5
pytest-cov
pylint
black
pre-commit

setuptools~=59.6.0
notanorm~=2.0.2
